Fixes bugs with deleting users, prevents admin from deleting himself

+++ cc-admin/members.php
@@ -24,10 +24,13 @@ if (!empty ($_GET['delete']) && is_numeric ($_GET['delete'])) {
 
     // Validate id
     $user = $userMapper->getUserById($_GET['delete']);
-    if ($user) {
+    if ($user && $user->userId != $adminUser->userId) {
         $userService->delete($user);
         $message = 'Member has been deleted';
         $message_type = 'success';
+    } else if ($user->userId == $adminUser->userId) {
+        $message = 'You can\'t delete yourself, silly!';
+        $message_type = 'errors';
     }
 
 }
